Kernel+Lib/Keyboard: add common keyboard functions in kernel_user_link
We have created a library that allow easy usage of keyboard
functionalities.

It can be used next by processes to easily get keyboard access
