+++ kernel/src/io/keyboard.rs
@@ -5,6 +5,7 @@ use core::{
 
 use alloc::sync::Arc;
 use blinkcast::alloc::{Receiver as BlinkcastReceiver, Sender as BlinkcastSender};
+pub use kernel_user_link::keyboard::{modifier, Key, KeyType};
 
 use crate::{
     cpu::{
@@ -51,268 +52,77 @@ const KEYBOARD_INT_NUM: u8 = 1;
 const KEYBOARD_STATUS_PORT: u16 = 0x64;
 const KEYBOARD_DATA_PORT: u16 = 0x60;
 
-/// The index is `KeyType`
-const US_KEYTYPE_KEYMAP: [u8; 127] = [
-    0, 27, b'1', b'2', b'3', b'4', b'5', b'6', b'7', b'8', b'9', b'0', b'-', b'=', b'\x08', b'\t',
-    b'q', b'w', b'e', b'r', b't', b'y', b'u', b'i', b'o', b'p', b'[', b']', b'\n', 0, b'a', b's',
-    b'd', b'f', b'g', b'h', b'j', b'k', b'l', b';', b'\'', b'`', 0, b'\\', b'z', b'x', b'c', b'v',
-    b'b', b'n', b'm', b',', b'.', b'/', 0, b'*', 0, b' ', 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
-    b'7', b'8', b'9', b'-', b'4', b'5', b'6', b'+', b'1', b'2', b'3', b'0', b'.', 0, 0, 0, 0, 0, 0,
-    0, b'\n', 0, 0, 0, 0, 0, 0, 0, 0, b'/', 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
-    0, 0, 0, 0, 0, 0, 0, 0,
-];
+// 0x80 means extended key
+fn key_type_from_device(value: u8) -> Option<KeyType> {
+    if value & 0x80 == 0 {
+        if value <= KeyType::F12 as u8 {
+            // not extended.
+            // we know that we are mapping the not extended keys directly
+            // so we can just cast it
+            let k = unsafe { core::mem::transmute(value) };
+            if k == KeyType::_None1
+                || k == KeyType::_None2
+                || k == KeyType::_None3
+                || k == KeyType::_None4
+            {
+                return None;
+            }
 
-/// The index is `KeyType`
-const US_KEYTYPE_KEYMAP_SHIFTED: [u8; 127] = [
-    0, 27, b'!', b'@', b'#', b'$', b'%', b'^', b'&', b'*', b'(', b')', b'_', b'+', b'\x08', b'\t',
-    b'Q', b'W', b'E', b'R', b'T', b'Y', b'U', b'I', b'O', b'P', b'{', b'}', b'\n', 0, b'A', b'S',
-    b'D', b'F', b'G', b'H', b'J', b'K', b'L', b':', b'"', b'~', 0, b'|', b'Z', b'X', b'C', b'V',
-    b'B', b'N', b'M', b'<', b'>', b'?', 0, b'*', 0, b' ', 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
-    0, 0, b'-', 0, b'5', 0, b'+', 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
-    0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
-];
-
-#[allow(dead_code)]
-#[repr(u8)]
-#[derive(Debug, Clone, Copy, PartialEq)]
-pub enum KeyType {
-    // normal keys (mapped 1:1 with set 1 scan codes)
-    _None1,
-    Escape,
-    Num1,
-    Num2,
-    Num3,
-    Num4,
-    Num5,
-    Num6,
-    Num7,
-    Num8,
-    Num9,
-    Num0,
-    Minus,
-    Equals,
-    Backspace,
-    Tab,
-    Q,
-    W,
-    E,
-    R,
-    T,
-    Y,
-    U,
-    I,
-    O,
-    P,
-    LeftBracket,
-    RightBracket,
-    Enter,
-    LeftCtrl,
-    A,
-    S,
-    D,
-    F,
-    G,
-    H,
-    J,
-    K,
-    L,
-    Semicolon,
-    SingleQuote,
-    Backtick,
-    LeftShift,
-    Backslash,
-    Z,
-    X,
-    C,
-    V,
-    B,
-    N,
-    M,
-    Comma,
-    Dot,
-    Slash,
-    RightShift,
-    KeypadAsterisk,
-    LeftAlt,
-    Space,
-    CapsLock,
-    F1,
-    F2,
-    F3,
-    F4,
-    F5,
-    F6,
-    F7,
-    F8,
-    F9,
-    F10,
-    NumLock,
-    ScrollLock,
-    Keypad7,
-    Keypad8,
-    Keypad9,
-    KeypadMinus,
-    Keypad4,
-    Keypad5,
-    Keypad6,
-    KeypadPlus,
-    Keypad1,
-    Keypad2,
-    Keypad3,
-    Keypad0,
-    KeypadDot,
-    _None2,
-    _None3,
-    _None4,
-    F11,
-    F12,
-
-    // extended keys
-    MultimediaPreviousTrack,
-    MultimediaNextTrack,
-    KeypadEnter,
-    RightCtrl,
-    MultimediaMute,
-    Calculator,
-    MultimediaPlayPause,
-    MultimediaStop,
-    VolumeDown,
-    VolumeUp,
-    WWWHome,
-    KeypadSlash,
-    RightAlt,
-    Home,
-    UpArrow,
-    PageUp,
-    LeftArrow,
-    RightArrow,
-    End,
-    DownArrow,
-    PageDown,
-    Insert,
-    Delete,
-    LeftGUI,
-    RightGUI,
-    Application,
-    Power,
-    Sleep,
-    Wake,
-    WWWSearch,
-    WWWFavorites,
-    WWWRefresh,
-    WWWStop,
-    WWWForward,
-    WWWBack,
-    MyComputer,
-    Email,
-    MultimediaSelect,
-}
-
-impl KeyType {
-    pub fn virtual_key(&self, shifted: bool) -> Option<u8> {
-        let index = *self as usize;
-        let mappings = if shifted {
-            &US_KEYTYPE_KEYMAP_SHIFTED
+            Some(k)
         } else {
-            &US_KEYTYPE_KEYMAP
-        };
-
-        assert!(index < mappings.len());
-        let value = mappings[index];
-
-        if value == 0 {
+            // not a valid key
             None
-        } else {
-            Some(value)
         }
-    }
-}
+    } else {
+        // first, strip the extension
+        let key = value & !0x80;
 
-impl TryFrom<u8> for KeyType {
-    type Error = ();
-
-    // 0x80 means extended key
-    fn try_from(value: u8) -> Result<Self, Self::Error> {
-        if value & 0x80 == 0 {
-            if value <= Self::F12 as u8 {
-                // not extended.
-                // we know that we are mapping the not extended keys directly
-                // so we can just cast it
-                let k = unsafe { core::mem::transmute(value) };
-                if k == Self::_None1 || k == Self::_None2 || k == Self::_None3 || k == Self::_None4
-                {
-                    return Err(());
-                }
-
-                Ok(k)
-            } else {
-                // not a valid key
-                Err(())
-            }
-        } else {
-            // first, strip the extension
-            let key = value & !0x80;
-
-            // use match normally
-            match key {
-                0x10 => Ok(Self::MultimediaPreviousTrack),
-                0x19 => Ok(Self::MultimediaNextTrack),
-                0x1C => Ok(Self::KeypadEnter),
-                0x1D => Ok(Self::RightCtrl),
-                0x20 => Ok(Self::MultimediaMute),
-                0x21 => Ok(Self::Calculator),
-                0x22 => Ok(Self::MultimediaPlayPause),
-                0x24 => Ok(Self::MultimediaStop),
-                0x2E => Ok(Self::VolumeDown),
-                0x30 => Ok(Self::VolumeUp),
-                0x32 => Ok(Self::WWWHome),
-                0x35 => Ok(Self::KeypadSlash),
-                0x38 => Ok(Self::RightAlt),
-                0x47 => Ok(Self::Home),
-                0x48 => Ok(Self::UpArrow),
-                0x49 => Ok(Self::PageUp),
-                0x4B => Ok(Self::LeftArrow),
-                0x4D => Ok(Self::RightArrow),
-                0x4F => Ok(Self::End),
-                0x50 => Ok(Self::DownArrow),
-                0x51 => Ok(Self::PageDown),
-                0x52 => Ok(Self::Insert),
-                0x53 => Ok(Self::Delete),
-                0x5B => Ok(Self::LeftGUI),
-                0x5C => Ok(Self::RightGUI),
-                0x5D => Ok(Self::Application),
-                0x5E => Ok(Self::Power),
-                0x5F => Ok(Self::Sleep),
-                0x63 => Ok(Self::Wake),
-                0x65 => Ok(Self::WWWSearch),
-                0x66 => Ok(Self::WWWFavorites),
-                0x67 => Ok(Self::WWWRefresh),
-                0x68 => Ok(Self::WWWStop),
-                0x69 => Ok(Self::WWWForward),
-                0x6A => Ok(Self::WWWBack),
-                0x6B => Ok(Self::MyComputer),
-                0x6C => Ok(Self::Email),
-                0x6D => Ok(Self::MultimediaSelect),
-                _ => Err(()),
-            }
+        // use match normally
+        match key {
+            0x10 => Some(KeyType::MultimediaPreviousTrack),
+            0x19 => Some(KeyType::MultimediaNextTrack),
+            0x1C => Some(KeyType::KeypadEnter),
+            0x1D => Some(KeyType::RightCtrl),
+            0x20 => Some(KeyType::MultimediaMute),
+            0x21 => Some(KeyType::Calculator),
+            0x22 => Some(KeyType::MultimediaPlayPause),
+            0x24 => Some(KeyType::MultimediaStop),
+            0x2E => Some(KeyType::VolumeDown),
+            0x30 => Some(KeyType::VolumeUp),
+            0x32 => Some(KeyType::WWWHome),
+            0x35 => Some(KeyType::KeypadSlash),
+            0x38 => Some(KeyType::RightAlt),
+            0x47 => Some(KeyType::Home),
+            0x48 => Some(KeyType::UpArrow),
+            0x49 => Some(KeyType::PageUp),
+            0x4B => Some(KeyType::LeftArrow),
+            0x4D => Some(KeyType::RightArrow),
+            0x4F => Some(KeyType::End),
+            0x50 => Some(KeyType::DownArrow),
+            0x51 => Some(KeyType::PageDown),
+            0x52 => Some(KeyType::Insert),
+            0x53 => Some(KeyType::Delete),
+            0x5B => Some(KeyType::LeftGUI),
+            0x5C => Some(KeyType::RightGUI),
+            0x5D => Some(KeyType::Application),
+            0x5E => Some(KeyType::Power),
+            0x5F => Some(KeyType::Sleep),
+            0x63 => Some(KeyType::Wake),
+            0x65 => Some(KeyType::WWWSearch),
+            0x66 => Some(KeyType::WWWFavorites),
+            0x67 => Some(KeyType::WWWRefresh),
+            0x68 => Some(KeyType::WWWStop),
+            0x69 => Some(KeyType::WWWForward),
+            0x6A => Some(KeyType::WWWBack),
+            0x6B => Some(KeyType::MyComputer),
+            0x6C => Some(KeyType::Email),
+            0x6D => Some(KeyType::MultimediaSelect),
+            _ => None,
         }
     }
 }
 
 #[allow(dead_code)]
-mod modifier {
-    pub const SHIFT: u8 = 1 << 0;
-    pub const CTRL: u8 = 1 << 1;
-    pub const ALT: u8 = 1 << 2;
-
-    pub const CAPS_LOCK: u8 = 1 << 3;
-    pub const NUM_LOCK: u8 = 1 << 4;
-    pub const SCROLL_LOCK: u8 = 1 << 5;
-    pub const EXTENDED: u8 = 1 << 6;
-
-    // a way to compress the `bytes` value
-    pub const PRESSED: u8 = 1 << 7;
-}
 
 const fn get_modifier(key: u8) -> Option<u8> {
     match key {
@@ -345,48 +155,6 @@ mod status {
     pub const KEYBOARD_TIMEOUT_MOUSE_DATA: u8 = 1 << 5;
     pub const RECEIVE_TIMEOUT: u8 = 1 << 6;
     pub const PARITY_ERROR: u8 = 1 << 7;
-}
-
-#[derive(Debug, Clone, Copy)]
-pub struct Key {
-    pub pressed: bool,
-    pub modifiers: u8,
-    pub key_type: KeyType,
-}
-
-impl Key {
-    pub const BYTES_SIZE: usize = 2;
-
-    /// # Safety
-    /// The `bytes` must be a valid representation of a `Key`
-    /// that has been created by `as_bytes`
-    #[allow(dead_code)]
-    pub unsafe fn from_bytes(bytes: [u8; Self::BYTES_SIZE]) -> Self {
-        let pressed = bytes[0] & modifier::PRESSED == 0;
-        let modifiers = bytes[0] & !modifier::PRESSED;
-        // Safety: we know that the `bytes` is a valid representation of `KeyType`
-        //         responsability of the caller to ensure that
-        let key_type = core::mem::transmute(bytes[1]);
-
-        Self {
-            pressed,
-            modifiers,
-            key_type,
-        }
-    }
-
-    pub fn as_bytes(&self) -> [u8; 2] {
-        let mut bytes = [0; 2];
-        bytes[0] = (self.modifiers & !modifier::PRESSED)
-            | if self.pressed { modifier::PRESSED } else { 0 };
-        bytes[1] = self.key_type as u8;
-        bytes
-    }
-
-    pub fn virtual_char(&self) -> Option<u8> {
-        let shifted = self.modifiers & modifier::SHIFT != 0;
-        self.key_type.virtual_key(shifted)
-    }
 }
 
 pub type KeyboardReader = BlinkcastReceiver<Key>;
@@ -440,7 +208,7 @@ impl Keyboard {
             // this is an extended key
             let data = self.read_data();
             let pressed = data & KEY_PRESSED == 0;
-            let key = (data | 0x80).try_into().ok()?;
+            let key = key_type_from_device(data | 0x80)?;
 
             return Some(Key {
                 pressed,
@@ -477,7 +245,7 @@ impl Keyboard {
             }
         }
         // this is a normal key
-        let key_type: KeyType = data.try_into().ok()?;
+        let key_type = key_type_from_device(data)?;
 
         Some(Key {
             pressed,
