+++ libraries/kernel_user_link/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "emerald_kernel_user_link"
-version = "0.2.4"
+version = "0.2.5"
 edition = "2021"
 readme = "README.md"
 authors = ["Amjad Alsharafi"]
